chore: Rename Custom Native Module on Android Env
